requirements.txt
Flask==2.2.5
flask-sqlalchemy==3.0.5
WTForms==3.0.1
flask-marshmallow==0.15.0
requests==2.31.0
beautifulsoup4==4.12.3
marshmallow-sqlalchemy==0.29.0
obs-websocket-py-1.0
Flask-APScheduler-1.12.4
Flask-SocketIO-5.3.6
